// types.ts
export interface RpmFormData {
  namaSatuanPendidikan: string;
  namaGuru: string;
  nipGuru: string;
  namaKepalaSekolah: string;
  nipKepalaSekolah: string;
  jenjang: string;
  kelas: string;
  mapel: string;
  capaianPembelajaran: string;
  materi: string;
  jumlahPertemuan: number;
  durasi: string;
  praktikPedagogis: string[];
  dimensiLulusan: string[];
}